Add vsmap arrayget/set, bump minor api version

+++ src/core/vscore.cpp
@@ -151,7 +151,7 @@ VSVariant::~VSVariant() {
     }
 }
 
-int VSVariant::size() const {
+size_t VSVariant::size() const {
     return internalSize;
 }
